requirements.txt
playwright>=1.46
pytest>=8.2
pyyaml>=6.0.2
loguru>=0.7.2
pillow>=10.4.0
imagehash>=4.3.1
numpy>=1.26.4
fakeredis>=2.3.0
redis>=5.0.0
pytest-xdist>=3.4.0 
pytest-html>=4.0.0
pytest-rerunfailures>=10.3 
opencv-python-headless>=    
scikit-image>=0.21.0    
